requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
snowflake-connector-python==3.6.0
sse-starlette==1.8.2
pydantic==2.5.3
